Switch to signal handler ML thread before checking invariant for mutator
An ML thread records the bytes requested at the time that it entered
the runtime.  If the runtime switches ML threads, then it must do so
before checking the heap invariant for the mutator, else it may not
leave sufficient bytes available for the resumed thread.

The switchToSignalHandlerThreadIfNonAtomicAndSignalPending uses
atomicState == 1 to detect if the current ML thread is in a non-atomic
context.  Normally, atomicState == 0 corresponds to a non-atomic
context, but switchToSignalHandlerThreadIfNonAtomicAndSignalPending
assumes that it as been called after an enter has bumped the
atomicState.