Treat stacks in uninitialized secondary threads as inactive.
In initWorld (init-world.c), called from GC_lateInit (init.c), the
primary pthread is allocated a new ML thread, which starts off with an
empty stack, and that ML thread/stack becomes the current.  The primary
pthread then starts the trampoline and begins executing the ML code,
which immediately allocates a frame, so that by the time of the first
entry to the runtime this stack is non-empty.

In duplicateWorld (init-world.c), called from GC_duplicate (init.c),
each of the secondary pthreads is allocated a new ML thread, which
starts off with an empty stack, and that ML thread/stack become the
current for the pthread.  Of course, these secondary pthreads do not
immediately start the trampoline and begin executing ML code.  Rather,
they wait for parallel initialization and then enter ML code through the
_export-ed Parallel_run. In switching to the C call handler, this ML
thread/stack is replaced by the C call handler (and saved in the C call
handler local data).

However, if a major GC occurs before the parallel initialization is
completed, then these "dummy" ML threads/stacks will be roots of the
secondary threads.  That old "current &&= not isStackEmpty(stack);" test
would catch this scenario.

In sequential MLton, other than that initially empty stack described
above before starting the trampoline, there should never be a live empty
stack --- every subsequent stack is always obtained by copying some
existing stack.  So, that justifies dropping the
"current &&= not isStackEmpty(stack);" test in mainline.

It might be better to initialize the secondary pthreads with
BOGUS_OBJPTR for their current thread, but there are various places
during GC (such as loop here that determines if a stack is a current
active stack) that assume that a current stack can be fetched from every
pthread.