Record procNumber in gcState
Record the processor number of each pthread during initialization,
rather than indirectly determining processor number via a search through
the procStates array.

In addition to being somewhat simpler, this works around a gcc
bug (Fedora 22, gcc 5.3.1) where all pthreads were acting like the
primary thread.  It seems that gcc is is miscompiling
MLton_threadFunc (in include/c-main.h) or Proc_amPrimary (in
runtime/gc/proc.c).  In particular, Proc_amPrimary is returning a bool
and the assembly code for Proc_amPrimary is writing a boolean (0 or 1)
result into %al (the low 8bits of %rax), but the upper bits are
non-zero, since %rax was used for a pointer value elsewhere in
Proc_amPrimary.  Meanwhile, MLton_threadFunc is using the result of
Proc_amPrimary via "testl %eax,%eax; je .Lnnn" (a fairly standard way of
teting a value for equality with 0).  This ends up always taking the
"true" branch, because the upper 24 bits of %eax are non-zero.