Distinguish size of header from size of metadata
Make a distinction in the runtime system between the header field of an
object and the entire metadata of an object.  For Normal, Stack, and
Weak objects, the metadata of an object is just the header field.  For
Array objects, the metadata of an object is the counter field, the
length field and the header field.

There is no functional difference with this commit, but it should
provide better infrastructure for future extensions of object metadata.