Add a tracing variant of the runtime system
The tracing runtime stores traces of runtime events on disk in an
unstable binary format for performance. The plan is to convert these
files into a more readily analyzable format and use them in debugging
and profiling tools for concurrent code.

Call MLton with "-trace-runtime true" to enable tracing. The traces
are stored in the directory pointed to by the MLTON_TRACE_DIR
environment variable. The directory must exist. If MLTON_TRACE_DIR is
not set, traces are not generated. For example:

  $ mlton-parmem -trace-runtime true prog.mlb
  $ MLTON_TRACE_DIR=/tmp ./prog
  $ ls /tmp/prog*.trace