runtime: Do HH Local Collection while queue lock is held
Previously, when the queue lock was held by the calling processor,
local collection was deferred. However, the logic assumed that the
collection happened even though it hadn't, leading to bad LCHS
resizes.

It's actually possible to collect while the calling processor holds
the queue lock, we just have to make sure not to unlock it on
collection completion. This commit enables that.