basis-library: Implement intermediate thread MLton.Thread.switch
This commit addresses issue #16. For all public atomicSwitch and
switch calls, the thread module will now go through a intermediate
thread to call the function passed to switch, avoiding the race
condition where another pthread switches to a MLton thread that has
not yet been switched out of, resulting in two pthreads running the
same MLton thread.

The implementation also implements exception handling from the
function passed to switch, but this feature has not been tested.

The new implementation includes a new guarantee that the function
passed to switch is evaluated in the global heap. Surprinslgy, this
actually makes the calls to switch in the hierarchical-heap scheduler
a bit easier to reason about.