Termination: add checks in busy loops
The new termination protocol is based on each processor periodically
checking whether a new termination request has been issued. Before
this commit, this check was only performed when entering the runtime
system, since I was assuming each pthread would periodically enter it.
However, this is not true in all cases. Consider the following:

1/ A tries to acquire lock X and succeeds.
2/ B tries to acquire lock X but has to wait for A.
3/ A starts a termination request, still holding X. It signals B and all
   other threads, and then waits for them to terminate.

We have reached a deadlock, since A is waiting for B and vice-versa.

One might consider that the problem here is that A should release the
lock before trying to terminate. That might be true for the runtime C
code, but not of arbitrary ML code: an ML program should be able to
terminate even if it holds a particular resource.

The solution we use here is to make the runtime primitives called from
the ML code "termination-aware". This means for instance that the C
primitives for acquiring locks in parallel.c have to check for
termination in their busy loops. Same for the busy initialization
loops in processor.c, which are indirectly called by the various
schedulers.

This commit adds the required code. In practice, this fixes a deadlock
in Spoonhower's scheduler arising for the reason outlined above, as
well as systematic deadlocks when running a sequential program with
number-processors > 1.