basis-library/schedulers/hierarchical-heap: Fix Entanglement issue
The job token (which is used to remove the job in the second half of
ForkJoin.fork) used to be a ref to a record containing refs of the
queue details. This use of a ref could cause down-pointer
entanglement.

The token was setup this way because in Spoonhower's scheduler, jobs
could move queues. However, we neither need nor support such behavior
anymore. As such, tokens are now just simple int pairs that are
generated per-processor. Also, removeWork now just check's the calling
processor's deque.

To elaborate why this is sufficient, consider that
1) Work stealing ensures that if a removeWork call fails, then all
ancestor queued tasks are also removed (and hence their corresponding
removeWork calls will fail).
2) Threads will only migrate processors due to suspension (which is
only caused by a failed removeWork call) .