basis-library/mlton/HM.sml: Remove broken linear HH implementation
Previously, I used a ref to invalidate HH's on setReturnValue (which
is effectively a cast). However, this doesn't actually work since it
only invalidates the HH passed in, not all references to it. In
addition, using refs is fraught with the HH scheduler.

As such, I removed this "feature" entirely. We should consider a
proper linear types/invalidation scheme in the future.