b-l/schedulers/hierarchical-heap: Finesse use of refs in syncvars
The syncvar now has separate state and value refs. In addition, fork
now stores the HH in the syncvar rather than a result datatype. This
seems to solve the entanglement issue we had previously. That being
said, this is further evidence that we need a better language or IL to
write the scheduler in.