runtime: Fix entanglement bug due to fragmentation optimization.
Our fragmentation optimization (i.e. where you allocate in an
ancestor's chunk) can cause entanglement on superheap merge. This can
happen when the parent superheap's allocation is occurring at an
ancestor of the merged superheap, leading to a pointer from the
ancestor chunk to the return value of the merged child superheap at a
lower depth.

This commit fixes this bug by enforcing that the parent superheap is
allocating at the current depth before merging in the child
superheap.