(WIP: deprecating ChunkPool_find)
Hackily handle large stacks in the same way as large arrays: by allocating a
new chunk if a stack crosses a block boundary.

Note that this strategy is bad because the new chunk could be immediately thrown
away. When the new block allocator is installed, this should be fixed by doing
one of the following.
  1. Set the chunk limit to force a failed limit check immediately after
     an allocation which crosses a block boundary. (After returning from a
     runtime, is the mutator permitted to allocate anything without doing a
     limit check first? Do we need to guarantee some number of bytes free upon
     return? Is this what thread->bytesNeeded is used for?)
  2. Allocate objects which would cross the block boundary separately (in their
     own fresh chunk).