Avoid setting HH for intermediate threads
Tweak the scheduler to use the "top-level" scheduler threads for allocating
new threads in child heaps. Previously, the intermediate threads (in the
"double-switch" trick used in the MLton.Thread library) were set to use a
child HH, which is difficult to reason about.

A secondary consequence of this change is that the promotion code needs to
be able to handle the case where either a scheduler or intermediate thread
performs a mutation, since the MLton.Thread library uses references.

NOTE: this change is largely experimental. One takeaway from this exploration
is that the scheduler threads are extremely tricky to reason about, and we
probably need to refactor the scheduler. It might be beneficial for the
scheduler to operate on primitive threads directly, rather than going through
the MLton.Thread library, since primitive threads are "closer" to the runtime.