Newly mmap'ed chunks at the front of the freelist
My intuition is that this should make finding a free chunk faster, since
we start at the beginning of the freelist and search forward until we find
a chunk that has enough space. Fresh chunks are large (controlled by
alloc-chunk) and therefore should be the first to be checked for free space
on subsequent allocations.