Single object chunks
Mark certain chunks as containing only a single object; then in a GC, these
chunks may simply be "moved" rather than copying their contents. This is
implemented by unlinking them from their current list, which has subtle
correctness requirements: it is incorrect to unlink a chunk
which is not a leaf in the path-compressing levelHead tree. At the moment we
guarantee this never happens by enforcing that single-object chunks are never
levelHeads, since if a chunk is not a leaf in the levelHead tree, then it must
have been a levelHead in the past.