Don't leak freed chunks in proc-local freelist
Chunks which are freed after a local GC are now put back in the local freelist,
to prevent a memory leak. Note however that the current policy is retained,
which only considers the first chunk in the freelist before mmap'ing more
space. So, these chunks are effectively still leaked at the moment. In the
future we can work on reusing these chunks.