Two per-proc local freelists
Each processor maintains two freelists of chunks, one of which is guaranteed
to have exactly one large chunk (for large allocations), and the other which
is used for small chunks. When a new chunk is needed, first the small chunk-list
is checked, then the large chunk. When chunks are freed, they are put in the
small chunk-list. We check at most a constant number of small chunks before
satisfying the request with the large chunk.