refactor HH structure: array instead of linked list
Previously the levels of an HH were organized into a linked list, which had
the advantage of not having a size limit. However, this prevents quickly
looking up the level of a known depth. It also unnecessarily complicated the
code in a number of places.

By using an array of levels instead, we get constant-time retrieval of a level.
This will become crucial in the implementation of remembered sets and deferred
promotion.

For now, the implementation uses a fixed capacity level array (64), but this
could be changed in the future to permit dynamic resizing. Note that for
parallel programs with reasonable span and granularity control, the fork depth
shouldn't get too big, so this should be sufficient for now.