Relaxed GC policy with ancestor heap capacities
The old GC policy reset the LCHS upon merge with a heuristic based on
the amount of newly unfrozen data in the ancestor. This policy effectively
caused the hierarchy to "forget" how big ancestor heaps are permitted to
become. For benchmarks with lots of data at the root, this was triggering
unnecessary root collections.

This patch introduces a strictly more relaxed policy which explicitly
remembers LCHS values at steals and restores them after a merge, if they are
larger than what the previous policy would have chosen.

Specifically we have three heaps: _P_arent, _L_eft child, and _R_ight child.
After the merge, the size and capacity of the resulting heap are set in terms
of the sizes and capacities of P, L, and R.

OLD POLICY:
  size = size(L) + size(R) + size(P)
  cap = cap(L) + cap(R) + 2*size(P)

NEW POLICY:
  size = size(L) + size(R) + size(P)
  cap = max(cap(P), cap(L) + cap(R) + 2*size(P))