bugfix: subtle frontier caching
Fixed a bug where, after growing a stack, the frontier of the
containing chunk was not being updated before switching to another chunk,
leaving frontiers in an incoherent state.

*Why this is so subtle*:
The runtime and mutator agree upon a frontier via s->frontier. The
runtime also has each chunk remember its frontier. Sometimes, these values
go out of sync, for performance reasons analogous to caching: if we just
operate on s->frontier, then we only need to update chunk->frontier when we
switch chunks. This leads to a frontier coherence problem: when is it guaranteed
that chunk->frontier has the most up-to-date value? When is it relied upon?

In this case I had forgotten that growStackCurrent does not ensure frontier
coherence.