Fix limit checks to guarantee chunk invariants
Previously, it was possible for a limit check to request more bytes than
can be accommodated by a single block of a chunk. This violates the chunk
invariant, that all object headers lie within the first block of their
constituent chunk.

This patch adjusts the pass for inserting limit checks:
  1. We split blocks of large allocations into multiple blocks of small
     allocations, to guarantee the chunk invariant for limit checks at the
     front of individual blocks.
  2. We permit coalescing limit checks (as before), but additionally insert
     limit checks to guarantee that each individual check is valid.