Fix write barriers with new Address operand
Extends RSSA and Machine operands with new form, Address, which
conceptually takes the "address-of" an operand. This is useful
because Offset and ArrayOffset operands in rvalue position
implicitly execute a dereference; taking their address explicitly
prevents this.

We use Address to express internal pointers, which are passed
to the runtime write barrier (as an objptr*) to indicate where
within an object there is an update. Address should never be used
to construct an internal pointer that the program "holds on to".
Also, we should only ever take the Address of Offset and ArrayOffset
operands.