Mostly collectible scheduler data
Each scheduler thread now owns a heap in the hierarchy which is an
immediate child of the root heap. The root heap holds items such as
mailboxes and scheduler deques, as well as any other data allocated
before the scheduler is initialized.

The only objects which are specially handled now are thread objects,
which are still explicitly allocated in the global heap.