Allocate threads in their own heap
Implicitly allocate new threads with their own heap, and use this space to
allocate the thread itself.

preThreads still don't have a hierarchicalHeap, but actual threads do.
I.e. when we copyCurrentThread we don't create a heap, but when we
copyThread we do (so preThreads are still just normal "ML data", whereas
runnable threads have a heap).