Simpler GC policy
A simpler GC policy avoids needing to explicitly keep track of the
local scope by instead preferring to use the size of the entire
superheap as its "current size". This size can be computed as needed,
at local GC checks, by adding up the size of ancestor levels.
Collections are triggered when the current size exceeds a collection
threshold which is adjusted after each local GC.

This simple policy has similar properties to the previous policy. For
example the amount of data promoted, which was explicitly handled in
the prior policy, is implicitly handled here. The "capacity" trick of
the previous policy is also essentially handled here by using the
entire superheap size to adjust collection thresholds.

The disadvantage of this policy is that it is somewhat asymmetric. One
of the workers will always have all of the "past data" in its superheap,
and will prefer fewer collections. In the future we plan to handle
"past data" separately from local collections, which should fix this issue.