Bugfixes
In hh collection, the existing algorithm was incorrect, but we must
have been getting lucky in the past. It is *not* the case that, after
promotion, every object has at most one forwarding pointer, because
promotions contained within the local scope will install a forwarding
pointer, and then a subsequent copy into the to-space will install
a second. So, we need to follow forwarding pointers until we find an
object that has none. (Technically this could be specialized since we
know that there can be at most 2 forwarding pointers, but a loop is
simpler.)

In the scheduler, we need to set the queue depth both before and after
the heap merge at the join point. Previously, if a GC was triggered
between setting the queue depth and doing the merge, it would be
incorrect (the unmerged heaps would be ignored).

I also properly initialized the queueTop in the GC_state, and handle the
case where it has not yet been registered by the scheduler.