Bugfixes for deque locking
A local GC needs to appear like a no-op to the mutator, but it also needs
to lock the deque. So now we use two different markers to indicate whether
or not the deque was locked by *this* pthread, or another pthread. The
GC should not wait for the deque to be unlocked when it was locked by
the same pthread, and must leave the lock state the same to how it was
when the GC began.

Also, this patch fixes the mistaken omission of locking on pushFront.