Fix ABP deque setDepth
When changing the depth of the deque, depending on whether or not the
desired depth is bigger or smaller than the current index, we have to
set the top and bot "in the correct order". This is because in the ABP
deque, it's safe for the "bottom" pointer to be less or equal to the
"top" pointer, but not the other way around unless there are items in
the deque to be stolen.