Heap refactor: linked list instead of preallocated levels
This patch replaces the array-of-levels hierarchy data structure
with a list-of-levels data structure.

Recall that each leaf task stores a leaf-to-root path of heaps.
Previously, these were stored in an array which is nice for fast
lookup but comes at a cost: we waste allocating space for unused
levels, and we are limited to a predetermined choice of a number
of levels. By using a list-of-levels, we only need to allocate
for the levels that we need, and the number of levels is unbounded.