Refactor handling of uninhabited types in `SimplifyTypes`
Following the previous commit, further use a `Type.t option` and
`NONE` to signal an uninhabited type, rather than a `Type.t` and
`Type.unit`.

In particular, a `Statement.t` that purports to produce a value of an
uninhabited type must be unreachable.  This optimization isn't
expressible when an uninhabited type is signaled by `Type.unit`,
because many statements may produce a value of `Type.unit`.