Fix bug in `SimplifyTypes`
After identifying an unreachable statement (e.g., the `ConApp` of a
`Useless` constructor), `SimplifyTypes` stops visiting subsequent
statements in the block.  However, visiting statements is necessary to
`setVarInfo` on the variables defined in the block.  Because
`SimplifyTypes` visits the blocks of a function via a DFS, it may
visit (necessarily unreachable) blocks that use variables that were
defined (but not visited) by a previously unreachable block.  This
triggers a `no varInfo property` compiler error.

This was previously noted (see the comment at the end of
MLton/mlton@19b07c0) and has now been observed in the wild
(see MPLLang/mpl#107).

To fix the bug, switch from a DFS to a dominator tree traversal, where
the dominator tree traversal does not visit any children of an
unreachable block.  This addresses the `no varInfo property` compiler
error, because any block that uses a variable must be dominated by the
block that defines the variable.