Further optimize `SimplifyTypes`
It is impossible for a primitive to be called with a value of an
uninhabited type.  Hence, a reachable primitive must have all
inhabited (and therefore useful) arguments.  As a further consequence,
a reachable `MLton_eq` and `MLton_equal` must have all inhabited (and
therefore useful) arguments and there will be no opportunity to
simplify the equality to `true`.