Further optimize `SimplifyTypes`
An array or vector of an uninhabited type is necessarily empty.

This is important, because while an `Array_sub` or `Vector_sub` of a
sequence of an uninhabited type will be recognized as unreachable, the
"origin" of the unreachability is the dominating bounds check that
must necessarily fail.  Replacing `Array_length` and `Vector_length`
of a sequence of an uninhabited type allows the dominating bounds
check to be simplified.