Pinned chunk-lists during collection. Pinning works!!
I switched to using explicit pinned chunk lists during collection. Now we
have a weaker invariant: chunks with pinned objects are generally unscannable.

Previously, I was trying to keep a stronger invariant, that all chunks should
always be scannable at collection. The idea was to fill the gaps in pinned
chunks with "dummy" objects that are safe to scan. But this was
too tricky to implement, and would not be as efficient overall.