Mutator header masking
MLton generates code that inspects headers to implement case expressions
more efficiently (by casing on the GC type of the object, which has to be
stored anyway for GC, we can determine which branch of an algebraic datatype
we are inspecting). MLton implements this by comparing the entire header
word against an expected static header, with the expectation that any
changes the runtime makes to headers will not be visible to mutator code.
See the genCase function in mlton/backend/packed-representation.fun.

With pinning, we use bits in the header to mark pinned objects. To make sure
that the mutator remains oblivious to such changes, we need to mask out
these bits in the generated code. This patch implements that change.

Here is an example program that fails without this patch.

```
(* An "interesting" datatype, where MLton can generate more efficient
 * case expressions based on inspecting headers *)
datatype foo = String of string | Chars of char list

val n = 1000 * 1000
val a: foo array = ForkJoin.alloc n

(* If called within a par, this will allocate an object and pin it *)
fun put j =
  Array.update (a, j-1,
    if j mod 3 = 0 then
      String (Int.toString j)
    else
      Chars (String.explode (Int.toString j)))

fun doit b i j =
  if j-i = 0 then b
  else if j-i = 1 then

    (* This cases on pinned objects in `a`. *)
    case Array.sub (a, i) of
      String s => doit (b + String.size s) (i+1) j
    | Chars cs => doit (b + List.length cs) (i+1) j

  else
    let
      val mid = i + (j-i) div 2

      (* This allocates an object and pins it by writing it into `a`,
       * which creates a down-pointer. *)
      val _ = put mid

      val (x, y) = ForkJoin.par (fn _ => doit b i mid, fn _ => doit b mid j)
    in
      x + y
    end

val r = (put n; doit 0 0 n)
val _ = print (Int.toString r ^ "\n")
```