Fix scheduler threads depths
There was a bug in initialization for setting up the scheduler
threads in the actual hierarchy, where some scheduler threads
were at depth 1, and other were at detpth 0. This fixes that bug.

Recall, the intended architecture is below, where H is the root
of the heap hierarchy (from the perspective of the source program)
and S0, S1, ... are the heaps of the scheduler threads, up to
(N-1) for N processors. The "global" heap, which is the only heap
at depth 0, contains any data allocated before scheduler
initialization, as well as shared scheduler data (e.g. deques)
and new thread records.

      +--------------------------+
   0  |         global           |
      +--------------------------+
        |      |       |      |
      +---+  +----+  +----+
   1  | H |  | S0 |  | S1 |  ...
      +---+  +----+  +----+

The cool thing about this architecture is that the existing local
collection algorithm works for collecting scheduler data S0, S1 etc.
without any modifications.

Originally, we tried accomplishing this by setting (in Parallel_init)
the depths of the other threads to 1. And actually, there was redundancy
because the initThreadAndHeap calls (init-world.c) allocated extra threads
at depth 1. However, due to other usage of threads in the MLton libraries
(e.g. Thread.register, used for FFI), the "current thread" of the
additional processors at the time of Parallel_init was not necessarily the
same physical thread as the one that will be used for the scheduler.

To fix this, we set initial thread depths to 0 and wait until we know
for sure that we are using a scheduler thread to set the depth to 1.