Enable scheduler thread collections
The existing GC policy disables local collections for shallow
depths, which also implicitly was preventing scheduler collections.
To fix this, I gave each thread a minLocalCollectionDepth
field which is the minimum depth permitted to be in-scope for
a local collection. By default, this is set according to the GC
policy, but can be overridden explicitly. The scheduler threads
set their own minLocalCollectionDepth to permit collections.