Seam carving example
Seam carving is an algorithm for "content-aware rescaling", allowing an
image to be rescaled while somewhat preserving the interesting features
of the image. It works by removing low-energy "seams" (vertical or
horizontal paths of pixels) to reduce either the width or height of
the image. For more info, see https://en.wikipedia.org/wiki/Seam_carving.

This patch provides a parallel implementation. It can remove as many
seams as desired, and can also generate an animation (as a .gif) of
the process, showing how the image shrinks gradually as more
seams are removed. It's super cool... check it out!

Implementing seam carving efficiently turned out to be quite interesting.
It's not as parallel as it seems (seams? haha). This is in part because
removing seams is entirely sequential: we can't remove the Nth seam
until we have removed the (N-1)th seam. But removing a single seam should
be highly parallel though, right? Not exactly. Let's see some of the
details...

At the center of the algorithm is a dynamic-programming equation of
the following form. (For vertical seams; horizontal are symmetric).

  M(i,j) = E(i,j) + min(M(i-1, j-1), M(i-1, j), M(i-1, j+1))

This equation at first appears highly parallel, because for example
each row of the image can be processed entirely in parallel. However,
in practice, it is difficult to extract this parallelism on multicore,
because a typical image will only be at most a few thousand pixels
wide. With proper granularity control of say up to 1000 pixels as
a unit, there's not much parallelism left!

So, row-major DP order isn't going to work. But this is not the only
possible ordering. Here's a better one. Imagine breaking up the image
into a bunch of strips, where each strip is then divided into triangles.
If each triangle is processed sequentially row-wise from top to bottom,
then we can compute a strip in parallel by first doing all of upper
triangles (#), and then doing all of the lower triangles (.):

             +------------------------------------+
  strip 1 -> |\####/\####/\####/\####/\####/\####/|
             |.\##/..\##/..\##/..\##/..\##/..\##/.|
             |..\/....\/....\/....\/....\/....\/..|
  strip 2 -> |\####/\####/\####/\####/\####/\####/|
             |.\##/..\##/..\##/..\##/..\##/..\##/.|
             |..\/....\/....\/....\/....\/....\/..|
  strip 3 -> |                                    |

Now, we can choose an appropriate grain size for a single triangle and
still manage to fit a decent number of triangles horizontally.
At the end of the day, however, this only gets us so far...