(WIP) decheck localized synch depths: avoid lookup on joins
This patch has two components:
  1. Some initial progress on proc-local synch depths. Just
  brainstorming for the most part. These parts are currently
  commented out.
  2. A slight change to how thread IDs are handled at
  joins, described below. This is important for the upcoming
  proc-local synch depths strategy, which eliminates the
  global lookup data structure for fork paths.

Explaining part 2
-----------------

In `GC_HH_decheckJoin`, previously there were two lookups into the
synch_depths for the left and right thread IDs. This was necessary
because "old" thread IDs were being passed as argument.

For example, in the following, at the t5 join point, the arguments
used to be t1 and t2. **With this change, the arguments to the t5
join point will be t3 and t4, instead of t1 and t2.**

```
      t0
    /    \
  t1      t2
 /  \    /  \
 *  *    *  *
 \  /    \  /
  t3      t4
    \    /
      t5
```

This change makes it possible to avoid doing synch_depth queries at
joins. Instead, we just use the recorded dag-depths in the argument
IDs.