decheck thread-local sync depths
Eliminate the global synch_depths lookup data structure, replacing
it with thread-local root-to-leaf paths. The synch_depths array
is now kept around only in the debug version of the runtime, to
double check that the thread-local sync depths are correct.

NOTE: Sync depths are THREAD-LOCAL. Previously the idea was to do
processor-local sync depths, but this doesn't work! (I previously
thought that sync-depths are similar to scheduler deque management,
which is why I assumed we could do proc-local. But deques are special:
we know that when a thread blocks at a join, it's deque is empty, so
it is safe to reuse the proc-local deque for the next thread. For
the sync depths, when a thread blocks at a join, its local sync depths
will still be needed in the continuation!)

This implementation works, but is slow and needs to be optimized.
In particular:
  1. Currently, LCA queries are expensive. The previous LCA mask
  bit tricks were buggy (I think). I replaced these with an
  obviously-correct-but-slow implementation.
  2. At steals, we copy the entire root-to-leaf path of thread-local
  sync depths. Since we currently have a small max fork depth in
  the checker anyway, this isn't a problem. But if we eventually
  switch to permitting unbounded fork-depths, this will become
  a problem.
  3. (Related to previous) when copying thread-local sync depths
  at steals, I thought we could just copy the shared ancestors. But
  this turned out to not be correct. Perhaps an off-by-one error?
  To work around this, I just copy the whole thing for now.
  See the TODO note in `runtime/gc/decheck.c` function
  `GC_HH_copySyncDepthsFromThread`.