Reverb example: much faster comb filter
Make the reverb example much faster with a better comb filter
algorithm. The previous implementation was as much as 200% slower
than the obvious sequential comb filter. This patch improves that
to only something like 50%.

Here's the basic idea. We lay out the input as a matrix (row-major
order). In this layout, each column is a distinct instance of a
"geometric prefix sums" problem (my terminology):

```
OUTPUT[i] = \sum_{j=0}^i \alpha^{i-j} * INPUT[j]
```

These "geometric prefix sums" can be computed with an algorithm
similar to blocked parallel prefix sums. For improved locality,
we also do multiple adjacent columns together as one unit. In this
way, the grains of the computation become square tiles of the input
matrix.

Implementing it is index-arithmetic-hell, but fairly straightforward.