Fix subtle race at join points
We use an incounter to synchronize two processors at a join point.
In the case where the right-hand processor arrives first at the
join first, it will decrement the incounter, observe that the other
processor hasn't completed yet (because the incounter is still
non-zero), and so it will call `GC_switchToThread` in order to switch
back to a scheduler thread. This call updates a little bit of metadata
on the thread that is being left behind (e.g. current chunk frontier
and current chunk list used-size) before performing the actual switch.

However, this is racy! As soon as the incounter is decremented, the
left-hand processor might arrive, decrement the incounter, and then
perform the thread merge. Merging threads is NOT safe for concurrency.

To fix this, I put a spin-loop into the thread-merge which waits for
the `GC_switchToThread` to mark the thread as not currently being
used. (Note that a similar spin-loop is already being used when we
switch to a thread, for similar reasons! It was an oversight to
not include it on the merge, too.)

The spin-loop is a bit unfortunate, but doesn't seem to have a big
impact on performance, because the "critical section" between the
incounter synchronization and the resulting switch is tiny -- just
a few instructions.

**Note for the future**: a better solution might be to first switch
away from the thread and then perform the incounter decrement.