stay at depth 2 below root CC until it completes
A step towards CC chaining. Previously, at the join point for
a root CC, the join always completes regardless of whether or
not the CC is done, causing the program to return to depth 1
even while the CC continues.

In this patch, the program stays at depth 2 until the CC
completes. At each fork, we check if the parent CC has completed,
and then merge upwards if it has. This works, but doesn't have
good performance, because CCs aren't chained yet: as the program
stays at depth 2, following CCs are handled by the processor on
the left-hand-side, causing (in the worst case) sequentialization
due to delaying parallelism as we wait for a CC at depth 2 to
complete.

Some specific changes:
  * `registerCont` now returns true/false indicating whether or
  not the heap is ready for CC. When forking CC, we now only do
  so if the heap is ready.
  * A new function `checkFinishedCCReadyToJoin` allows us to
  delay the join until a CC has completed. At the moment this
  is only valid for root CC, but in the future we'll generalize.