Better CC chaining policy and incremental merging
Introduced new `hh->subHeapCompletedCC` which is a region that
holds regions for completed CCs, all merged together. When checking
if a new CC should be triggered, we first find all completed CCs
and merge them into this region. This guarantees that we don't
repeatedly pay for traversing the CC chain.

We don't actually merge the `subHeapCompletedCC` into the main heap
until the CC-chain is empty. This guarantees that the snapshotting
method is still correct.

We now trigger new CCs when the _combined size_ of the completed and
fresh region get too big. This ensures that we don't trigger too
many CCs, and can allow old CCs to catch up, to transition to a
more accurate root-set snapshot for the combined region.