Handle remembered down-pointers into CCed regions
This change is necessary in anticipation of permitting CCed
subheaps throughout the hierarchy (not just at the root).

Each hh has a remembered set which stores objects that are
referenced by down-pointers. These remembered sets are not
concurrent data-structures! When we detect a down-pointer into
a CC'ed subregion, where that CC is being performed by another
processor, we cannot just add the new entry into the remembered
set.

To fix this, we use the special "subHeapCompletedCC" subregion
to store down-pointer entries into regions that are being CCed.
All subheaps in a cc-chain hold a pointer to this region, and
as they complete are incrementally merged in. Once the cc-chain
has been completely merged in, we finally merge the
subHeapCompletedCC into the main heap, at which point the
remembered set is valid for a new collection.