any-level CC-chaining seems to be working
The main change needed to get it to work was restricting the scope
of local collections, to ensure that local collections don't
overlap with CC. Specifically, if a level has at least one
registered or in-progress CC, then we have to exclude it from local
collection.

We might be able to relax this restriction if we edit local collection
to update rootsets for the CC. But this is really tricky. Much easier
to just avoid the issue entirely. It doesn't seem like it would have
a huge impact on performance.