Fix race in the CC chain
In-progress CCs might do heap queries on same-level or ancestor
objects. For this to be safe, we need to guarantee that the heaps
they query are not being concurrently merged with other heaps.
Unfortunately my design for eagerly merging completed CCs into
the hh->subHeapCompletedCC was causing such a race.

To fix it, rather than having a single hh->subHeapCompletedCC which
accumulates completed regions, this patch puts completed CCs into a
secondary chain.

The new design looks like the following, where <<-- indicates a
subHeapForCC link, and <-- indicates a subHeapCompletedCC link.

                 ... <<-- CCed heap <<-- MAIN_HEAP
                  |             |          |
                  +-----------+ |          |
                              | |          |
                              v v          |
                             primary       |
      ... <-- completed <-- completed <----+

There is a special "primary" completed heap at the front of the
completed chain, which is used to store new remset entries for
CCed subheaps, as before. All CCed subheaps contain a pointer to
this primary completed subheap, which is used in the write barrier.
The change in this patch is that the completed subheaps are now
linked into a chain, rather than being eagerly merged into the
primary completed subheap.