Fix liveness for pcall
There was an error where the parr continuation had a different
frame size than the parl/cont continuations.

Actually, the previous liveness calculation allowed for both
parl/parr to have different sizes than the cont. This occurs
because the size of the frame is calculated based off the
maximum live slot, but the liveness is different in the
different continuations.

This fix is a bit of a hack: we just force the liveness of
all three to match, which guarantees that the size of the
frames will match.

One potentially bad thing: this can keep data live longer
than necessary (e.g. if the parr completes, then all data
needed *only by* parr is kept live on the parl
continuation, and this won't be collected until we complete
the first parl block). This... doesn't seem that bad. But
we should investigate.