Bugfix: exception handling and CGC scheduling
This fixes a subtle bug where, when scheduling a CGC, an unhandled
exception can skip the cleanup code that happens after checking if
the CGC-task can be popped and executed locally.

In the scheduler, in `forkGC`, there is a call to `fork ... (f, g)`
between the `push` and `popDiscard` for the CGC-task. Previously,
if this call raised an exception, then the `popDiscard` (and other
associated cleanup) was skipped.

The solution is straightforward, similar to how exceptions are
normally caught and re-raised across task boundaries:
  1. catch the exception if needed (via `result (fn () => ...)`)
  2. perform the cleanup
  3. re-raise the exception if needed (via `extractResult`)